Backport Finder and Config classes from 2.x line
