+++ Symfony/CS/Console/Command/FixCommand.php
@@ -187,12 +187,12 @@ to the default list of symfony-level fixers:
 
     <?php
 
-    \$finder = Symfony\CS\Finder\DefaultFinder::create()
+    \$finder = Symfony\CS\Finder::create()
         ->exclude('somedir')
         ->in(__DIR__)
     ;
 
-    return Symfony\CS\Config\Config::create()
+    return Symfony\CS\Config::create()
         ->fixers(array('strict_param', 'short_array_syntax'))
         ->finder(\$finder)
     ;
@@ -204,11 +204,11 @@ then specify all fixers to be used:
 
     <?php
 
-    \$finder = Symfony\CS\Finder\DefaultFinder::create()
+    \$finder = Symfony\CS\Finder::create()
         ->in(__DIR__)
     ;
 
-    return Symfony\CS\Config\Config::create()
+    return Symfony\CS\Config::create()
         ->level(Symfony\CS\FixerInterface::NONE_LEVEL)
         ->fixers(array('trailing_spaces', 'encoding'))
         ->finder(\$finder)
@@ -222,12 +222,12 @@ Note the additional <comment>-</comment> in front of the Fixer name.
 
     <?php
 
-    \$finder = Symfony\CS\Finder\DefaultFinder::create()
+    \$finder = Symfony\CS\Finder::create()
         ->exclude('somedir')
         ->in(__DIR__)
     ;
 
-    return Symfony\CS\Config\Config::create()
+    return Symfony\CS\Config::create()
         ->fixers(array('-psr0'))
         ->finder(\$finder)
     ;
@@ -238,7 +238,7 @@ The ``symfony`` level is set by default, you can also change the default level:
 
     <?php
 
-    return Symfony\CS\Config\Config::create()
+    return Symfony\CS\Config::create()
         ->level(Symfony\CS\FixerInterface::PSR2_LEVEL)
     ;
 
@@ -263,7 +263,7 @@ speed up further runs.
 
     <?php
 
-    return Symfony\CS\Config\Config::create()
+    return Symfony\CS\Config::create()
         ->setUsingCache(true)
     ;
 
@@ -340,8 +340,8 @@ EOF
         } elseif (file_exists($configFile)) {
             $config = include $configFile;
             // verify that the config has an instance of Config
-            if (!$config instanceof Config) {
-                throw new InvalidConfigurationException(sprintf('The config file "%s" does not return a "Symfony\CS\Config\Config" instance. Got: "%s".', $configFile, is_object($config) ? get_class($config) : gettype($config)));
+            if (!$config instanceof ConfigInterface) {
+                throw new InvalidConfigurationException(sprintf('The config file "%s" does not return a "Symfony\CS\ConfigInterface" instance. Got: "%s".', $configFile, is_object($config) ? get_class($config) : gettype($config)));
             }
 
             if ('txt' === $input->getOption('format')) {
