+++ Symfony/CS/Finder/DefaultFinder.php
@@ -11,63 +11,26 @@
 
 namespace Symfony\CS\Finder;
 
-use Symfony\Component\Finder\Finder;
-use Symfony\CS\FinderInterface;
+use Symfony\CS\Finder as BaseFinder;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @deprecated
  */
-class DefaultFinder extends Finder implements FinderInterface
+class DefaultFinder extends BaseFinder
 {
     public function __construct()
     {
+        @trigger_error(
+            sprintf(
+                'The "%s" class is deprecated. You should stop using it, as it will soon be removed in 2.0 version. Use "%s" instead.',
+                __CLASS__,
+                'Symfony\CS\Finder'
+            ),
+            E_USER_DEPRECATED
+        );
+
         parent::__construct();
-
-        $files = $this->getFilesToExclude();
-
-        $this
-            ->files()
-            ->name('*.php')
-            ->name('*.twig')
-            ->name('*.xml')
-            ->name('*.yml')
-            ->ignoreDotFiles(true)
-            ->ignoreVCS(true)
-            ->exclude('vendor')
-            ->filter(
-                function (\SplFileInfo $file) use ($files) {
-                    return !in_array($file->getRelativePathname(), $files, true);
-                }
-            )
-        ;
-    }
-
-    public function setDir($dir)
-    {
-        $this->in($this->getDirs($dir));
-    }
-
-    /**
-     * Gets the directories that needs to be scanned for files to validate.
-     *
-     * @param string $dir
-     *
-     * @return string[]
-     */
-    protected function getDirs($dir)
-    {
-        return array($dir);
-    }
-
-    /**
-     * Excludes files because modifying them would break.
-     *
-     * This is mainly useful for fixtures in unit tests.
-     *
-     * @return string[]
-     */
-    protected function getFilesToExclude()
-    {
-        return array();
     }
 }
