+++ Symfony/CS/Finder/MagentoFinder.php
@@ -13,13 +13,19 @@ namespace Symfony\CS\Finder;
 
 /**
  * @author Myke Hines <myke@webhines.com>
+ *
+ * @deprecated
  */
 class MagentoFinder extends DefaultFinder
 {
     public function __construct()
     {
         @trigger_error(
-            sprintf('The "%s" class is deprecated. You should stop using it, as it will soon be removed in 2.0 version.', __CLASS__),
+            sprintf(
+                'The "%s" class is deprecated. You should stop using it, as it will soon be removed in 2.0 version. Use "%s" instead.',
+                __CLASS__,
+                'Symfony\CS\Finder'
+            ),
             E_USER_DEPRECATED
         );
